V13: (no description)
Differential Revision: D15793003
Differential Version: 85209962

+++ aten/src/ATen/test/quantized_test.cpp
@@ -16,8 +16,8 @@ using namespace at;
 TEST(TestQTensor, QuantDequantAPIs) {
   auto num_elements = 10;
   Tensor r = at::ones({num_elements});
-  const float scale = 1.0;
-  const int32_t zero_point = 2;
+  const double scale = 1.0;
+  const int64_t zero_point = 2;
   Tensor qr = at::quantize_linear(r, scale, zero_point, kQUInt8);
   ASSERT_EQ(qr.q_scale(), scale);
   ASSERT_EQ(qr.q_zero_point(), zero_point);
